HSEARCH-3619 Move HibernateSearchContextService to the mapping package
This class will depend on almost everything in the ORM mapper, there's
no way around it. So let's put it at the heart of it, then work on
removing cyclic dependencies.
