+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/event/impl/OptimalEventsHibernateSearchState.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.mapper.orm.event.impl;
 
-import org.hibernate.search.mapper.orm.impl.HibernateSearchContextService;
+import org.hibernate.search.mapper.orm.mapping.impl.HibernateSearchContextService;
 import org.hibernate.search.util.common.impl.Contracts;
 
 /**
