+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/session/impl/SearchSessionWritePlanImpl.java
@@ -7,7 +7,7 @@
 package org.hibernate.search.mapper.orm.session.impl;
 
 import org.hibernate.engine.spi.SessionImplementor;
-import org.hibernate.search.mapper.orm.impl.HibernateSearchContextService;
+import org.hibernate.search.mapper.orm.mapping.impl.HibernateSearchContextService;
 import org.hibernate.search.mapper.orm.session.SearchSessionWritePlan;
 import org.hibernate.search.mapper.pojo.work.spi.PojoWorkPlan;
