+++ mapper/orm/src/main/resources/META-INF/services/org.hibernate.service.spi.ServiceContributor
@@ -1,1 +1,1 @@
-org.hibernate.search.mapper.orm.impl.HibernateSearchContextServiceContributor
+org.hibernate.search.mapper.orm.mapping.impl.HibernateSearchContextServiceContributor
